+++ lib/index.js
@@ -4,47 +4,28 @@ var assert = require('assert');
 var debug = require('debug')('thalisalti:acl');
 var fast = require('./indexOf');
 
-var noRole = 'public';
 var NOTFOUND = -1;
 var IDTOKEN = '{:id}';
 
 
 function lookupPathVerb(req, paths, lookupPath) {
-  debug('looking up: %s : %s', lookupPath);
   var pathExists = fast.indexOf(paths, lookupPath, 'path');
-  debug(pathExists);
-  if (NOTFOUND !== pathExists) {
-    //verbs are just an array of strings
-    var verbs = paths[pathExists].verbs;
-    debug('verbs: %s', JSON.stringify(verbs));
-    var verbExists = fast.indexOf(verbs, req.method);
-    debug(verbExists);
 
-    if (NOTFOUND === verbExists) {
-      debug('verb and path not found: req.Path: %s  lookupPath: %s', req.path, lookupPath);
-      return false;
-    }
-    return true;
-  }
-  else {
+  if (NOTFOUND === pathExists) {
     return false;
   }
+  //verbs are just an array of strings
+  var verbs = paths[pathExists].verbs;
+  var verbExists = fast.indexOf(verbs, req.method);
+
+  if (NOTFOUND === verbExists) {
+    debug('verb and path not found: req.Path: %s  lookupPath: %s', req.path, lookupPath);
+    return false;
+  }
+  return true;
+
 }
 
-/**
- * Acl
- * @typedef {Object} Salti.Acl
- * TODO:
- */
-
-/**
- * Implements an Express middleware Authentication Filter
- * that validates an identity provided as a property on the request.connection object
- * against ACL represented as a json object.
- * @param {String} dbname - dbname known to provider
- * @param {Salti.Acl} inAcl - acl represented as an array of path.role.verb
- * @returns {Function} returns a handler of (req, res, next) to the middleware pipeline
- */
 module.exports = function(dbname, inAcl) {
 
   if (!dbname || typeof dbname !== 'string') {
@@ -64,8 +45,12 @@ module.exports = function(dbname, inAcl) {
     var okPathVerb = false;
     debug('method: %s url: %s path: %s', req.method, req.url, req.path);
 
-    //using headers for validation of the acl engine       
-    var pskRole = req.connection.pskRole || noRole;
+    if (!req.connection.pskRole) {
+      debug('unauthorized0: no role: %s ', req.path);
+      return res.status(401).send(msg401);
+    }
+
+    var pskRole = req.connection.pskRole;
     debug('pskRole: %s', pskRole);
 
     var roleExists = fast.indexOf(acl, pskRole, 'role');
@@ -76,7 +61,7 @@ module.exports = function(dbname, inAcl) {
     }
 
     var paths = acl[roleExists].paths;
-    
+
     //here we're doing a strict simple RAW path lookup on ACLs
     var rawExists = fast.indexOf(paths, req.path, 'path');
 
@@ -96,7 +81,7 @@ module.exports = function(dbname, inAcl) {
     else {
       //this section deals with /db/id, /db/_local/id, and /db/id/attachment stuff.
       var pathParts = req.path.split('/');
-      
+
       if (pathParts[0] !== '') {
         //sanity check.. the first element should always be empty.
         debug('unauthorized4.0: %s : %s', pskRole, req.path);
@@ -104,7 +89,7 @@ module.exports = function(dbname, inAcl) {
       }
 
       var lookupPath = req.path.substring(0, req.path.lastIndexOf('/') + 1) + IDTOKEN;
-      
+
       if (pathParts.length === 3) {
         //we have just a path and ID
         //do a search on /db/id and bail if NO GOOD
@@ -117,8 +102,6 @@ module.exports = function(dbname, inAcl) {
         }
       }
       else if (pathParts.length === 4 && pathParts[2] === '_local') {
-        debug('pathParts.length: %s   pathParts[2]: %s', pathParts.length, pathParts[2]);
-        //it's either _local or id/attachment
         if (!lookupPathVerb(req, paths, lookupPath)) {
           debug('unauthorized4.2: req.Path: %s  req.path: %s', req.path, lookupPath);
           return res.status(401).send(msg401);
