+++ lib/indexOf.js
@@ -28,15 +28,10 @@
 
 /**
  * Custom indexOf implementation from fast.js.
- * @module indesOf
+ * @module indexOf
  */
 function compareIt(source, target) {
-    if (source === target) {
-      return true;
-    }
-    else {
-      return false;
-    }
+  return source === target;
 }
 
 /**
@@ -44,7 +39,6 @@ function compareIt(source, target) {
  * @param {Object} subject - the source array or object.
  * @param {string} target - the value to lookup.
  * @param {string} key - an optional property value to use if using an object as subject
- * @param {Number} fromIndex - where to start in the subject (offset)
  * @returns {Number} the index value where found or -1 if not found
  */
 module.exports.indexOf = function fastIndexOf(subject, target, key) {
