+++ sample/app/validate.js
@@ -33,12 +33,3 @@ function genDoc(){
   return JSON.stringify({ _id: guid(), "type": "foobar" })
 }
 
-function guid() {
-  function s4() {
-    return Math.floor((1 + Math.random()) * 0x10000)
-      .toString(16)
-      .substring(1);
-  }
-  return s4() + s4() + '-' + s4() + '-' + s4() + '-' +
-    s4() + '-' + s4() + s4() + s4();
-}
