+++ sample/routes/replicate.js
@@ -58,14 +58,4 @@ router.post('/add', function (req, res, next) {
 })
 
 
-function guid() {
-  function s4() {
-    return Math.floor((1 + Math.random()) * 0x10000)
-      .toString(16)
-      .substring(1);
-  }
-  return s4() + s4() + '-' + s4() + '-' + s4() + '-' +
-    s4() + '-' + s4() + s4() + s4();
-}
-
 module.exports = router;
