--- sample/runit.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-nodemon -i node_modules/
