--- test/acl-get-user.js
@@ -1,25 +1,0 @@
-'use strict';
-
-module.exports =  [
-  {
-    "path": '/',
-    "roles": [
-      {"role": 'user',
-        "verbs": ['GET', 'PUT', 'POST']}
-    ]
-  },
-  {
-    "path": '/foo',
-    "roles": [
-      {"role": 'user',
-        "verbs": ['GET', 'PUT', 'POST']}
-    ]
-  },
-  {
-    "path": '/bar',
-    "roles": [
-      {"role": 'user',
-        "verbs": ['GET']}
-    ]
-  }
-];
