--- test/acl-path-types.js
@@ -1,32 +1,0 @@
-'use strict';
-
-module.exports =  [
-  {
-    "path": '/base',
-    "roles": [
-      {"role": 'public',
-        "verbs": ['GET', 'PUT', 'POST']}
-    ]
-  },
-    {
-    "path": '/foo',
-    "roles": [
-      {"role": 'public',
-        "verbs": ['GET', 'PUT', 'POST']}
-    ]
-  },
-    {
-    "path": '/bar/foo',
-    "roles": [
-      {"role": 'public',
-        "verbs": ['GET', 'PUT', 'POST']}
-    ]
-  },
-    {
-    "path": '/fiz/baz',
-    "roles": [
-      {"role": 'public',
-        "verbs": ['GET', 'PUT', 'POST']}
-    ]
-  }
-];
