+++ test/test-noaclfile.js
@@ -2,11 +2,11 @@
 
 var request = require('supertest'),
   express = require('express'),
-  path = require('path'),
+  fspath = require('path'),
   colors = require('colors'),
   assert = require('assert');
 
-var lib = require(path.join(__dirname, '../lib/index'));
+var lib = require(fspath.join(__dirname, '../lib/index'));
 
 function genericHandlers(router) {
   var handlers = require('./handlers2');
@@ -16,12 +16,6 @@ function genericHandlers(router) {
   return router;
 }
 
-/**
- * we define 2 distinct 'roles' (other than Admin)
- * 1. Pull Replication
- * 2. Beacon  - using a well-known PSK identity of beacon
- * All other requests are non-authenticated and purely anonymous are rejected.
- */
 describe('test-noaclfile.js - should let all through', function() {
   describe(' - simple check with NO acl - no identity', function() {
     var app, router;
@@ -29,14 +23,10 @@ describe('test-noaclfile.js - should let all through', function() {
     router = express.Router();
 
     before(function() {
-      //mocker..
       router.all('*', function(req, res, next) {
-        //req.connection.pskIdentity = 'public';//not adding any identity for faking.
         next();
       })
-      //Norml middleware usage..
-      //router.all('*', lib([{}]));
-      //mock handlers  
+ 
       app.use('/', genericHandlers(router));
     })
     it('should be 200', function(done) {
@@ -53,14 +43,11 @@ describe('test-noaclfile.js - should let all through', function() {
     router = express.Router();
 
     before(function() {
-      //mocker..
       router.all('*', function(req, res, next) {
         req.connection.pskIdentity = 'public';
         next();
       })
-      //Norml middleware usage..
       router.all('*', lib('foobar', [{}]));
-      //mock handlers  
       app.use('/', genericHandlers(router));
     })
     it('should be 401', function(done) {
